fix(sonar): generic fix of bugs

+++ src/integration/lib/src/main/java/se/tink/backend/aggregation/eidassigner/QsealcSignerHttpClient.java
@@ -156,7 +156,7 @@ public class QsealcSignerHttpClient {
      * it.
      */
     public CloseableHttpResponse execute(HttpPost post) throws IOException {
-        for (int i = 1; i <= MAX_RETRY_ATTEMPTS; i++) {
+        for (long i = 1; i <= MAX_RETRY_ATTEMPTS; i++) {
             try {
                 return httpClient.execute(post);
             } catch (IOException e) {
